Modification cloud-servicechain-state.yang key
Due to elan-service-chain-state is an augmentation of
elan-instance, elan-instance-name can be removed from
elan-to-pseudo-port-data list key. The new key will be
composed with elanLportTag and scfTag.

ScfTag is now defined as an unsigned 32-bit number

+++ vpnservice/cloud-servicechain/cloud-servicechain-impl/src/main/java/org/opendaylight/netvirt/cloudservicechain/listeners/VpnPseudoPortListener.java
@@ -57,12 +57,12 @@ public class VpnPseudoPortListener
     @Override
     protected void update(InstanceIdentifier<VpnToPseudoPortTag> identifier, VpnToPseudoPortTag original,
                           VpnToPseudoPortTag update) {
-        VpnPseudoPortCache.addVpnPseudoPortToCache(update.getVrfId(), update.getLportTag());
+        VpnPseudoPortCache.addVpnPseudoPortToCache(update.getVrfId(), update.getLportTag().intValue());
     }
 
     @Override
     protected void add(InstanceIdentifier<VpnToPseudoPortTag> identifier, VpnToPseudoPortTag add) {
-        VpnPseudoPortCache.addVpnPseudoPortToCache(add.getVrfId(), add.getLportTag());
+        VpnPseudoPortCache.addVpnPseudoPortToCache(add.getVrfId(), add.getLportTag().intValue());
     }
 
     @Override
